--- src/__mocks__/hooks/useContentView.js
@@ -1,14 +1,0 @@
-export const mockGoTo = jest.fn();
-
-const useContentView = () => ({
-  displayedComponent: 'Home',
-  goTo: mockGoTo,
-  isMapVisible: false,
-  setIsMapVisible: jest.fn(),
-  navbarColor: 'blue',
-  setNavbarColor: jest.fn(),
-  navbarLeftButton: 'back',
-  setNavbarLeftButton: jest.fn(),
-});
-
-export default useContentView;
